+++ src/components/Resolution/index.tsx
@@ -22,6 +22,7 @@ interface Props {
   value?: number | null;
   onChange?: (num: number | null) => void;
   placeholder?: string;
+  width?: number | string;
 }
 
 const options = [10, 15, 30, 60, 120, 300].map((num) => ({
@@ -30,7 +31,7 @@ const options = [10, 15, 30, 60, 120, 300].map((num) => ({
 }));
 
 export default function Resolution(props: Props) {
-  const { onChange, value, placeholder } = props;
+  const { onChange, value, placeholder, width } = props;
   const [inputContent, setInputContent] = useState<string>(String(value || ''));
   const [setp, setStep] = useState<number | null>(value || null);
 
@@ -73,7 +74,7 @@ export default function Resolution(props: Props) {
       <AutoComplete
         options={options}
         value={inputContent}
-        style={{ width: 72 }}
+        style={{ width: width ?? 72 }}
         onChange={onInputChange}
         onSelect={handelSelect}
         onKeyDown={handleEnter}