+++ src/pages/dashboard/Renderer/Renderer/TimeSeriesNG/Main.tsx
@@ -29,7 +29,6 @@ import './style.less';
 export { getDataFrameAndBaseSeries };
 
 interface Props {
-  dashboardID: number;
   id: string;
   frames: AlignedData;
   baseSeries: BaseSeriesItem[];
@@ -56,7 +55,6 @@ export default function index(props: Props) {
   const history = useHistory();
   const location = useLocation();
   const {
-    dashboardID,
     frames,
     baseSeries,
     darkMode,
@@ -77,7 +75,7 @@ export default function index(props: Props) {
     onZoomWithoutDefult,
   } = props;
   const id = isPreview ? `${props.id}__view` : props.id;
-  const { custom, options = {}, targets, overrides } = panel;
+  const { custom, options = {}, targets, overrides, queryOptionsTime } = panel;
   const [dashboardMeta] = useGlobalState('dashboardMeta');
   const uplotRef = useRef<any>();
   // 保存 x 和 y 轴初始缩放范围
@@ -86,8 +84,8 @@ export default function index(props: Props) {
   const [showResetZoomBtn, setShowResetZoomBtn] = useState(false);
   const [annotationSettingUp, setAnnotationSettingUp] = useState(false);
   const xMinMax = useMemo(() => {
-    return getScalesXMinMax({ range, panel });
-  }, [range, JSON.stringify(_.map(panel.targets, 'time'))]);
+    return getScalesXMinMax({ range, queryOptionsTime });
+  }, [range, JSON.stringify(queryOptionsTime)]);
 
   const uOptions: Options = useMemo(() => {
     const yRange = getScalesYRange({ panel });
@@ -108,7 +106,6 @@ export default function index(props: Props) {
           renderFooter: (domNode: HTMLDivElement, closeOverlay: () => void) => {
             ReactDOM.render(
               <AddAnnotationButton
-                dashboardID={dashboardID}
                 panelID={id}
                 timeZone={timezone}
                 closeOverlay={closeOverlay}